wip loading for adding code sets
