+++ client/src/pages/Configurations/ConfigBuild/AddConditionCodeSets.tsx
@@ -5,17 +5,9 @@
  */
 import React, { useState } from 'react';
 import Drawer from '../../../components/Drawer';
-import ConditionCodeSetListItem from './ConditionCodeSet';
+import { ConditionCodeSetListItem } from './ConditionCodeSetListItem';
 import { Link } from 'react-router';
-import {
-  getGetConfigurationQueryKey,
-  useAssociateConditionWithConfiguration,
-  useDisassociateConditionWithConfiguration,
-} from '../../../api/configurations/configurations';
-import { useToast } from '../../../hooks/useToast';
-import { useQueryClient } from '@tanstack/react-query';
 import { highlightMatches } from '../../../utils/highlight';
-import { useApiErrorFormatter } from '../../../hooks/useErrorFormatter';
 import { IncludedCondition } from '../../../api/schemas';
 
 type AddConditionCodeSetsDrawerProps = {
@@ -42,65 +34,6 @@ const AddConditionCodeSetsDrawer: React.FC<AddConditionCodeSetsDrawerProps> = ({
         cond.display_name.toLowerCase().includes(searchTerm.toLowerCase())
       )
     : conditions;
-
-  const { mutate: associateMutation } =
-    useAssociateConditionWithConfiguration();
-  const { mutate: disassociateMutation } =
-    useDisassociateConditionWithConfiguration();
-
-  const showToast = useToast();
-  const queryClient = useQueryClient();
-  const formatError = useApiErrorFormatter();
-
-  // Add/remove handlers
-  function handleAssociate(conditionId: string) {
-    associateMutation(
-      {
-        configurationId,
-        data: { condition_id: conditionId },
-      },
-      {
-        onSuccess: async (resp) => {
-          showToast({
-            heading: 'Condition added',
-            body: resp.data.condition_name,
-          });
-          await queryClient.invalidateQueries({
-            queryKey: getGetConfigurationQueryKey(configurationId),
-          });
-        },
-      }
-    );
-  }
-
-  function handleDisassociate(conditionId: string) {
-    disassociateMutation(
-      {
-        configurationId,
-        conditionId,
-      },
-      {
-        onSuccess: async (resp) => {
-          showToast({
-            heading: 'Condition removed',
-            body: resp.data.condition_name,
-          });
-          await queryClient.invalidateQueries({
-            queryKey: getGetConfigurationQueryKey(configurationId),
-          });
-        },
-        onError: (error) => {
-          const errorDetail =
-            formatError(error) || error.message || 'Unknown error';
-          showToast({
-            variant: 'error',
-            heading: 'Error removing condition',
-            body: errorDetail,
-          });
-        },
-      }
-    );
-  }
 
   return (
     <Drawer
@@ -157,15 +90,8 @@ const AddConditionCodeSetsDrawer: React.FC<AddConditionCodeSetsDrawerProps> = ({
             return (
               <ConditionCodeSetListItem
                 key={key}
-                conditionName={condition.display_name}
-                associated={condition.associated}
+                condition={condition}
                 configurationId={configurationId}
-                onAssociate={() => {
-                  handleAssociate(condition.id);
-                }}
-                onDisassociate={() => {
-                  handleDisassociate(condition.id);
-                }}
                 highlight={highlight}
               />
             );
